統一: プロジェクト配下ページのヘッダーデザインを完全統一
## 主な変更内容

### ヘッダー構造の統一
- 全9ページ（プロジェクト詳細 + 8ワークフローページ）のPageHeaderを統一
- 04-question-groupページのレイアウトを `flex h-full flex-col` 構造に変更
- 次へボタンをフッター部分からPageHeader内に移動

### タイトル・説明文の統一
- 全ページの説明的なタイトルに変更（例：「採点」→「答案の採点・点数入力」）
- description プロパティを完全削除
- PageHeaderコンポーネントからdescription関連コード完全削除

### 次へボタンの統一
- 04-question-groupの青色ボタンスタイルを標準デザインに統一
- ボタンテキストをパンくずリスト表記と完全一致（例：「次へ: 7. 採点」）
- ワークフロー順序の修正（03→04→05の正しい遷移）

### 不要コンポーネント削除
- TemplateHeader.tsx（PageHeader直接使用に変更）
- ProjectStudentsPage.tsx（ページに直接実装）
- ProjectStudentsPageContainer.tsx（ページに直接実装）
- StudentAnswersPageHeader関数（PageHeader直接使用）

## 統一後の標準構造
```typescript
<div className="flex h-full flex-col">
  <PageHeader
    title="[説明的なタイトル]"
    helpButton={helpButton}
  >
    <Button onClick={nextStep}>次へ: [N]. [ステップ名]</Button>
  </PageHeader>
  <div className="flex-1 overflow-auto p-6">
    {/* メインコンテンツ */}
  </div>
</div>
```
