+++ components/layout/PageHeader.tsx
@@ -4,14 +4,12 @@ import React from "react"
 
 interface PageHeaderProps {
   title: string
-  description: string
   children?: React.ReactNode
   helpButton?: React.ReactNode
 }
 
 export default function PageHeader({
   title,
-  description,
   children,
   helpButton,
 }: PageHeaderProps) {
@@ -21,7 +19,6 @@ export default function PageHeader({
         <div className="flex items-center">
           <div>
             <h1 className="text-xl font-semibold">{title}</h1>
-            <p className="text-muted-foreground text-sm">{description}</p>
           </div>
           {helpButton && (
             <div className="ml-3">{helpButton}</div>
