--- components/projects/05-students/components/ProjectStudentsPage.tsx
@@ -1,7 +1,0 @@
-"use client"
-
-import { ProjectStudentsPageContainer } from "@/components/projects/05-students/components/project-students-page/components/ProjectStudentsPageContainer"
-
-export default function ProjectStudentsPage() {
-  return <ProjectStudentsPageContainer />
-}
